refactor(seo): v1.35.0 PR-F — split seo.js into seo/ directory-module
Split lib/commands/seo.js (1060 lines) into a seo/ directory-module mirroring
the icerik/ + mobile/ pattern. The last large-file split candidate.

- seo/_shared.js: shared HTML helpers (extractTag/Meta/AllMeta/Headings/Images/
  Links, stripTags, countWords), fetchPage, fetchDDG, parseDDGResults, SEO_UA.
- seo/{audit,meta,sitemap,speed,backlinks,rank,compare}.js (one per subcommand)
  + index.js (runSeo dispatch) + help.js. Verbatim move — CLI grammar + all
  output strings unchanged. parseDDGResults re-exported from index.js (tests
  import it); bin/badi.js + test import path updated to seo/index.js.
- help-doctor audits seo/ as a directory-module (still 38 commands drift-free).

DELIBERATELY NOT done: stripTags -> node-html-parser (the original P5 item).
Empirically a behavior regression — node-html-parser .text inserts a newline for
<br>/block elements ("<h1>a<br>b</h1>" -> "a\nb" vs regex "ab") and does not
replicate the fixpoint loop's nested-injection collapse. The security concern was
already resolved by the fixpoint loop (audit reclassified to style-only), so the
regex stays. Rationale recorded in CHANGELOG.

Tests 1317 (unchanged; pure move). biome clean. release check green. Closes the
v1.35.0 hardening sequence (A-F).
